% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar.R
\name{baixar_img_audio}
\alias{baixar_img_audio}
\title{Baixar uma imagem e um audio}
\usage{
baixar_img_audio(dir = "data-raw/captcha")
}
\arguments{
\item{dir}{diretório onde os arquivos serão salvos. O nome do arquivo é um
timestamp. A imagem e o áudio contém o mesmo CAPTCHA.}
}
\description{
Baixa uma imagem JPG e um audio WAV no caminho especificado.
}
