% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar.R
\name{baixar_imgs_audios}
\alias{baixar_imgs_audios}
\title{Baixar várias imagens e vários áudios}
\usage{
baixar_imgs_audios(nlim = 1000L, esperar = 0, dir = "data-raw/captcha",
  verbose = TRUE)
}
\arguments{
\item{nlim}{quantidade de captchas a serem baixados. Default 1000.}

\item{esperar}{quantidade de segundos para esperar entre requisições. Default
0.0.}

\item{dir}{diretório onde os arquivos serão salvos. O nome do arquivo é um
timestamp. Default "data-raw/captcha".}

\item{verbose}{imprimir mensagens com o andamento dos downloads?}
}
\description{
Baixa imagens JPG e áudios WAV no caminho especificado.
}
