% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar.R
\name{desenhar_corte}
\alias{desenhar_corte}
\title{Desenhar corte}
\usage{
desenhar_corte(onda)
}
\arguments{
\item{onda}{vetor de frequências retornado por \code{\link{ler_audio}}.}
}
\description{
Desenha uma onda após realizar o corte, pintando cada letra identificada com
uma cor diferente.
}
\examples{
arq_aud <- system.file('audio.wav', package = 'captchaTJMGAudio')
onda <- ler_audio(arq_aud)
desenhar_corte(onda)
}
