% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar.R
\name{desenhar_img}
\alias{desenhar_img}
\title{Desenha imagem}
\usage{
desenhar_img(arq_img)
}
\arguments{
\item{arq_img}{caminho da imagem jpg.}
}
\description{
Plota imagem na tela a partir de arquivo de imagem jpg.
}
