% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.R
\name{ler_audio}
\alias{ler_audio}
\title{Carrega o sinal .wav no R.}
\usage{
ler_audio(arq_aud)
}
\arguments{
\item{arq_aud}{caminho/do/audio/do/captcha.wav ou objeto do tipo response.}
}
\description{
Carrega o sinal .wav no R.
}
