% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predizer.R
\name{predizer}
\alias{predizer}
\title{Prediz as letras a partir do audio}
\usage{
predizer(arq_aud)
}
\arguments{
\item{arq_aud}{caminho/do/audio/do/captcha.wav ou objeto do tipo response.}
}
\description{
Recebe o caminho do áudio e retorna cinco números colados em um character
vector. Os números são obtidos a partir de uma tabela pré classificada das
somas das frequências de cada letra.
}
\examples{
arq_aud <- system.file('audio.wav', package = 'captchaTJMGAudio')
predizer(arq_aud)
# [1] "99182"
}
