% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picotar.R
\name{quebrar_letras}
\alias{quebrar_letras}
\title{Quebra onda em letras}
\usage{
quebrar_letras(onda)
}
\arguments{
\item{onda}{vetor de frequências retornado por \code{\link{ler_audio}}.}
}
\description{
Esta função faz as seguintes tarefas: 1. limpa o som, substituindo por zero
todas as frequências com valores absolutos menores que 2000. 2. Retira os
zeros iniciais e os zeros finais. 3. Identifica todos os intervalos contíguos
de zeros e seleciona os 4 maiores intervalos. Os cortes são definidos como o
ponto médio de cada intervalo contíguo de zeros.
}
