% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picotar.R
\name{tabelar_letras}
\alias{tabelar_letras}
\title{Monta tabela de letras e soma de frequências}
\usage{
tabelar_letras(arq_aud)
}
\arguments{
\item{arq_aud}{caminho/do/audio/do/captcha.wav ou objeto do tipo response.}
}
\description{
Essa função usa \code{\link{quebrar_letras}} para quebrar as letras do áudio.
Depois, agrupa por letra e soma todas as frequências. Como não há ruído, a
soma das frequências identifica unicamente o número e, portanto, basta
tabelar as somas, associando cada soma a um número manualmente.
}
\examples{
arq_aud <- system.file('audio.wav', package = 'captchaTJMGAudio')
tabelar_letras(arq_aud)
}
